requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
cycler==0.10.0
discord.py==1.7.3
idna==3.2
idna-ssl==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
multidict==5.1.0
numpy==1.19.5
Pillow==8.3.2
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
torch==1.4.0+cpu
torchvision==0.5.0+cpu
typing-extensions==3.10.0.2
yarl==1.6.3
